beater/api/asset/sourcemap: move decoder.ReqDecoder here
Move decoder.ReqDecoder to beater/api/asset/sourcemap,
which is the only remaining user. Also, move the decoder
implementation to the same package, which is the only user.
