+++ decoder/req_decoder.go
@@ -21,21 +21,15 @@ import (
 	"compress/gzip"
 	"compress/zlib"
 	"encoding/json"
-	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
-	"strings"
 
 	"github.com/pkg/errors"
 
 	"github.com/elastic/beats/libbeat/monitoring"
-
-	"github.com/elastic/apm-server/utility"
 )
 
 type ReqReader func(req *http.Request) (io.ReadCloser, error)
-type ReqDecoder func(req *http.Request) (map[string]interface{}, error)
 
 var (
 	decoderMetrics                = monitoring.Default.NewRegistry("apm-server.decoder", monitoring.PublishExpvar)
@@ -103,30 +97,4 @@ func DecodeJSONData(reader io.Reader) (map[string]interface{}, error) {
 		return nil, errors.Wrap(err, "data read error")
 	}
 	return v, nil
-}
-
-func DecodeSourcemapFormData(req *http.Request) (map[string]interface{}, error) {
-	contentType := req.Header.Get("Content-Type")
-	if !strings.Contains(contentType, "multipart/form-data") {
-		return nil, fmt.Errorf("invalid content type: %s", req.Header.Get("Content-Type"))
-	}
-
-	file, _, err := req.FormFile("sourcemap")
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	sourcemapBytes, err := ioutil.ReadAll(file)
-	if err != nil {
-		return nil, err
-	}
-	payload := map[string]interface{}{
-		"sourcemap":       string(sourcemapBytes),
-		"service_name":    req.FormValue("service_name"),
-		"service_version": req.FormValue("service_version"),
-		"bundle_filepath": utility.CleanUrlPath(req.FormValue("bundle_filepath")),
-	}
-
-	return payload, nil
 }
