+++ decoder/req_decoder_test.go
@@ -18,18 +18,13 @@
 package decoder_test
 
 import (
-	"bytes"
 	"encoding/json"
-	"io"
-	"mime/multipart"
-	"net/http"
 	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
 	"github.com/elastic/apm-server/decoder"
-	"github.com/elastic/apm-server/tests/loader"
 )
 
 func TestDecodeJSONData(t *testing.T) {
@@ -42,39 +37,4 @@ func TestDecodeJSONData(t *testing.T) {
 		"system": map[string]interface{}{"hostname": "prod1.example.com"},
 		"number": json.Number("123"),
 	}, decoded)
-}
-
-func TestDecodeSourcemapFormData(t *testing.T) {
-
-	body := &bytes.Buffer{}
-	writer := multipart.NewWriter(body)
-
-	fileBytes, err := loader.LoadDataAsBytes("../testdata/sourcemap/bundle.js.map")
-	assert.NoError(t, err)
-	part, err := writer.CreateFormFile("sourcemap", "bundle_no_mapping.js.map")
-	assert.NoError(t, err)
-	_, err = io.Copy(part, bytes.NewReader(fileBytes))
-	assert.NoError(t, err)
-
-	writer.WriteField("bundle_filepath", "js/./test/../bundle_no_mapping.js.map")
-	writer.WriteField("service_name", "My service")
-	writer.WriteField("service_version", "0.1")
-
-	err = writer.Close()
-	assert.NoError(t, err)
-
-	req, err := http.NewRequest("POST", "_", body)
-	req.Header.Set("Content-Type", writer.FormDataContentType())
-	assert.NoError(t, err)
-
-	assert.NoError(t, err)
-	data, err := decoder.DecodeSourcemapFormData(req)
-	assert.NoError(t, err)
-
-	assert.Len(t, data, 4)
-	assert.Equal(t, "js/bundle_no_mapping.js.map", data["bundle_filepath"])
-	assert.Equal(t, "My service", data["service_name"])
-	assert.Equal(t, "0.1", data["service_version"])
-	assert.NotNil(t, data["sourcemap"].(string))
-	assert.Equal(t, len(fileBytes), len(data["sourcemap"].(string)))
 }
